requirements.txt
streamlit
pyodbc
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
openai
